local/requirements.txt
flask
vosk
boto3
transformers
ffmpeg-python
google-api-python-client
python-dotenv
requests
urllib3